+++ scripts/build-26-targets.ps1
@@ -1,6 +1,7 @@
 param(
     [string[]]$Targets = @('26.1', '26.1.1', '26.1.2', '26.2'),
-    [switch]$SkipTests
+    [switch]$SkipTests,
+    [switch]$RequireCleanGit
 )
 
 $ErrorActionPreference = 'Stop'
@@ -9,6 +10,7 @@ $arguments = @{
     RootTargets = $Targets
     SkipClassic = $true
     SkipTests = $SkipTests.IsPresent
+    RequireCleanGit = $RequireCleanGit.IsPresent
 }
 
 & $AuditedBuilder @arguments